# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "asimaster",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/products/{product_id}/costs": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Get Cost Items",
        "operationId": "get_cost_items_api_v1_products__product_id__costs_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostItemResponse"
                  },
                  "title": "Response Get Cost Items Api V1 Products  Product Id  Costs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "costs"
        ],
        "summary": "Save Cost Items",
        "operationId": "save_cost_items_api_v1_products__product_id__costs_put",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CostItemCreate"
                },
                "title": "Items"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostItemResponse"
                  },
                  "title": "Response Save Cost Items Api V1 Products  Product Id  Costs Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/cost-presets": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Get Cost Presets",
        "operationId": "get_cost_presets_api_v1_users__user_id__cost_presets_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostPresetResponse"
                  },
                  "title": "Response Get Cost Presets Api V1 Users  User Id  Cost Presets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "costs"
        ],
        "summary": "Create Cost Preset",
        "operationId": "create_cost_preset_api_v1_users__user_id__cost_presets_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostPresetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostPresetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cost-presets/{preset_id}": {
      "put": {
        "tags": [
          "costs"
        ],
        "summary": "Update Cost Preset",
        "operationId": "update_cost_preset_api_v1_cost_presets__preset_id__put",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Preset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostPresetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostPresetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "costs"
        ],
        "summary": "Delete Cost Preset",
        "operationId": "delete_cost_preset_api_v1_cost_presets__preset_id__delete",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cost-presets/{preset_id}/apply": {
      "post": {
        "tags": [
          "costs"
        ],
        "summary": "Apply Cost Preset",
        "operationId": "apply_cost_preset_api_v1_cost_presets__preset_id__apply_post",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Preset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostPresetApplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostPresetApplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crawl/product/{product_id}": {
      "post": {
        "tags": [
          "crawl"
        ],
        "summary": "Crawl Product",
        "operationId": "crawl_product_api_v1_crawl_product__product_id__post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrawlKeywordResult"
                  },
                  "title": "Response Crawl Product Api V1 Crawl Product  Product Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawl/user/{user_id}": {
      "post": {
        "tags": [
          "crawl"
        ],
        "summary": "Crawl User",
        "operationId": "crawl_user_api_v1_crawl_user__user_id__post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlBatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawl/status/{user_id}": {
      "get": {
        "tags": [
          "crawl"
        ],
        "summary": "Get Crawl Status",
        "operationId": "get_crawl_status_api_v1_crawl_status__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawl/logs/{user_id}": {
      "get": {
        "tags": [
          "crawl"
        ],
        "summary": "Get Crawl Logs",
        "operationId": "get_crawl_logs_api_v1_crawl_logs__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrawlLogResponse"
                  },
                  "title": "Response Get Crawl Logs Api V1 Crawl Logs  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CostItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "type": {
            "type": "string",
            "pattern": "^(percent|fixed)$",
            "title": "Type"
          },
          "value": {
            "type": "number",
            "minimum": 0.0,
            "title": "Value"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "value"
        ],
        "title": "CostItemCreate"
      },
      "CostItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "name",
          "type",
          "value",
          "sort_order",
          "created_at"
        ],
        "title": "CostItemResponse"
      },
      "CostPresetApplyRequest": {
        "properties": {
          "product_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Product Ids"
          }
        },
        "type": "object",
        "required": [
          "product_ids"
        ],
        "title": "CostPresetApplyRequest"
      },
      "CostPresetApplyResponse": {
        "properties": {
          "applied": {
            "type": "integer",
            "title": "Applied"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "skipped_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Skipped Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "applied",
          "skipped"
        ],
        "title": "CostPresetApplyResponse"
      },
      "CostPresetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CostItemCreate"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "name",
          "items"
        ],
        "title": "CostPresetCreate"
      },
      "CostPresetResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CostItemCreate"
            },
            "type": "array",
            "title": "Items"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "items",
          "created_at"
        ],
        "title": "CostPresetResponse"
      },
      "CostPresetUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CostItemCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "title": "CostPresetUpdate"
      },
      "CrawlBatchResult": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "success": {
            "type": "integer",
            "title": "Success"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "total",
          "success",
          "failed"
        ],
        "title": "CrawlBatchResult"
      },
      "CrawlKeywordResult": {
        "properties": {
          "keyword_id": {
            "type": "integer",
            "title": "Keyword Id"
          },
          "keyword": {
            "type": "string",
            "title": "Keyword"
          },
          "items_count": {
            "type": "integer",
            "title": "Items Count"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "keyword_id",
          "keyword",
          "items_count",
          "success",
          "error"
        ],
        "title": "CrawlKeywordResult"
      },
      "CrawlLogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "keyword_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "keyword_id",
          "status",
          "error_message",
          "duration_ms",
          "created_at"
        ],
        "title": "CrawlLogResponse"
      },
      "CrawlStatusResponse": {
        "properties": {
          "total_keywords": {
            "type": "integer",
            "title": "Total Keywords"
          },
          "last_24h_success": {
            "type": "integer",
            "title": "Last 24H Success"
          },
          "last_24h_failed": {
            "type": "integer",
            "title": "Last 24H Failed"
          },
          "avg_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "total_keywords",
          "last_24h_success",
          "last_24h_failed",
          "avg_duration_ms"
        ],
        "title": "CrawlStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}